pyproject.toml
[project]
name = "commodity-registry"
version = "0.1.1"
description = "Single source of truth for financial instrument data"
authors = [
    { name = "Roman Medvedev", email = "pypi@romavm.dev" }
]
license = "MIT"
requires-python = ">=3.9"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
]
readme = "README.md"

[project.urls]
Homepage = "https://github.com/romamo/commodity-registry"
Repository = "https://github.com/romamo/commodity-registry"
Issues = "https://github.com/romamo/commodity-registry/issues"

[project.optional-dependencies]
fetchers = [
    "py-yfinance>=0.1.2",
    "py-ftmarkets>=0.1.2",
]

[project.scripts]
commodity-reg = "commodity_registry.cli:main"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.9.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/commodity_registry"]

[tool.hatch.build.targets.sdist]
include = ["src/commodity_registry"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
